requirements.txt
firebase_admin==6.6.0
Flask==3.1.0
Werkzeug==3.1.3
flask_apscheduler
rapidfuzz==2.4.3
flask-socketio==5.5.1
beautifulsoup4==4.13.3
python-rapidjson
gunicorn
python-dotenv
numpy
opencv-python
imutils
pillow
mediapipe